% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_1_iso}
\alias{baseline_1_iso}
\title{Stable isotope data for amphipods (baseline 1)}
\format{
\code{data.frame} containing 14 rows and 5 variables
\describe{
\item{common_name}{name of the spcies (i.e., Amphipoda)}
\item{ecoregion}{ecoregion where samples were collected}
\item{c2}{observed values for \eqn{\delta^{13}}C}
\item{n2}{observed values for \eqn{\delta^{15}}N}
}
}
\usage{
baseline_1_iso
}
\description{
Stable isotope data (\eqn{\delta^{13}}C and
\eqn{\delta^{15}}N) amphipods samples from an ecoregion in
Lake Ontario.
}
\keyword{datasets}
