% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_2_iso}
\alias{baseline_2_iso}
\title{Stable isotope data for lake trout, dreissenids (pelagic baseline), and
amphipods (benthic baseline) used in two source model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 5 columns.
}
\usage{
baseline_2_iso
}
\description{
Stable isotope data (SIA) for \eqn{\delta}\eqn{^1}\eqn{^3}C and
\eqn{\delta}\eqn{^1}\eqn{^5}N for lake trout collected from an ecoregion in
Lake Ontario. Values of \eqn{\delta}\eqn{^1}\eqn{^3}C and
\eqn{\delta}\eqn{^1}\eqn{^5}N for a pelagic baseline
(dreissenids; \code{c1} and \code{n1}) and a benthic baseline (amphipods;
\code{c2} and \code{n2}).
}
\keyword{datasets}
