% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_iso}
\alias{combined_iso}
\title{Stable isotope data for lake trout, dreissenids (pelagic baseline), and
amphipods (benthic baseline) used in two source model}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 13 columns.
}
\usage{
combined_iso
}
\description{
Stable isotope data (SIA) for \eqn{\delta^{13}}C and
\eqn{\delta ^{15}}N for lake trout collected from an ecoregion in
Lake Ontario. Values of \eqn{\delta ^{13}} C and
\eqn{\delta ^{1}}N for a pelagic baseline
(dreissenids; \code{c1} and \code{n1}) and a benthic baseline (amphipods;
\code{c2} and \code{n2}) with the means for each calcualted (\code{c1_mean}, \code{c2_mean},
\code{n1_mean}, \code{n2_mean}).
}
\keyword{datasets}
