% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consumer_iso}
\alias{consumer_iso}
\title{A \code{data.frame} of stable isotope data for lake trout used}
\format{
\code{data.frame} containing 30 rows and 6 variables
\describe{
\item{id}{row id number}
\item{common_name}{name of the spcies (i.e., Lake Trout)}
\item{ecoregion}{ecoregion where samples were collected}
\item{d13c}{observed values for \eqn{\delta^{13}C}
\item{d15n}{observed values for \eqn{\delta^{15}N}
}
}
\usage{
consumer_iso
}
\description{
Stable isotope data (SIA) for \eqn{\delta^{13}}C and
\eqn{\delta^ {15}}N for lake trout collected from an ecoregion in
Lake Ontario.
}
\keyword{datasets}
