% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_iso_osn}
\alias{example_iso_osn}
\title{Stable isotope data for lake trout and dreissenids (baseline) used in
a one source model}
\format{
\code{data.frame} containing 30 rows and 7 variables
\describe{
\item{common_name}{name of the spcies (i.e., Lake Trout)}
\item{ecoregion}{ecoregion where samples were collected}
\item{d13c}{observed values for \eqn{\delta}\eqn{^1}\eqn{^3}C}
\item{d15n}{observed values for \eqn{\delta}\eqn{^1}\eqn{^5}N}
\item{lambda}{trophic position of baseline used in this data (i.e., 2)}
\item{c1}{\eqn{\delta}\eqn{^1}\eqn{^3}C for baseline samples}
\item{n1}{\eqn{\delta}\eqn{^1}\eqn{^5}N for baseline samples}
}
}
\usage{
example_iso_osn
}
\description{
Stable isotope data (SIA) for \eqn{\delta}\eqn{^1}\eqn{^3}C and
\eqn{\delta}\eqn{^1}\eqn{^5}N for lake trout collected from an ecoregion in
Lake Ontario. The \eqn{\delta}\eqn{^1}\eqn{^3}C (\code{c1}) and
\eqn{\delta}\eqn{^1}\eqn{^5}N  (\code{n1}) values for dreissenid samples
from the same ecoregion.
}
\keyword{datasets}
