% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_iso_tsn}
\alias{example_iso_tsn}
\title{Stable isotope data for lake trout, dreissenids (pelagic baseline), and
amphipods (benthic baseline) used in two source model}
\format{
\code{data.frame} containing 87 rows and 10 variables
\describe{
\item{id}{row id number}
\item{common_name}{name of the spcies (i.e., Lake Trout)}
\item{ecoregion}{ecoregion where samples were collected}
\item{d13c}{observed values for \eqn{\delta}\eqn{^1}\eqn{^3}C}
\item{d15n}{observed values for \eqn{\delta}\eqn{^1}\eqn{^5}N}
\item{c1}{mean \eqn{\delta}\eqn{^1}\eqn{^3}C for pelagic baseline samples}
\item{n1}{mean \eqn{\delta}\eqn{^1}\eqn{^5}N for pelagic baseline samples}
\item{c2}{mean \eqn{\delta}\eqn{^1}\eqn{^3}C for benthic baseline samples}
\item{n2}{mean \eqn{\delta}\eqn{^1}\eqn{^5}N for benthic baseline samples}
\item{lambda}{trophic position of baseline used in this data (i.e., 2)}
}
}
\usage{
example_iso_tsn
}
\description{
Stable isotope data (SIA) for \eqn{\delta}\eqn{^1}\eqn{^3}C and
\eqn{\delta}\eqn{^1}\eqn{^5}N for lake trout collected from an ecoregion in
Lake Ontario. Values of \eqn{\delta}\eqn{^1}\eqn{^3}C and
\eqn{\delta}\eqn{^1}\eqn{^5}N for a pelagic baseline
(dreissenids; \code{c1} and \code{n1}) and a benthic baseline (amphipods;
\code{c2} and \code{n2}).
}
\keyword{datasets}
