% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_output_os}
\alias{model_output_os}
\title{Model Output - One Source Model
One source model output from \code{{brms}} model.}
\format{
\code{brms} object
}
\usage{
model_output_os
}
\description{
Model Output - One Source Model
One source model output from \code{{brms}} model.
}
\keyword{datasets}
