% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_output_os_mg}
\alias{model_output_os_mg}
\title{Model Output - One Source Model - Multiple Groups
One source model output from \code{{brms}} model for multiple groups.}
\format{
\code{brms} object
}
\usage{
model_output_os_mg
}
\description{
Model Output - One Source Model - Multiple Groups
One source model output from \code{{brms}} model for multiple groups.
}
\keyword{datasets}
