% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_output_ts}
\alias{model_output_ts}
\title{Model Output - Two Source Model
Two source model output from \code{{brms}} model. .}
\format{
\code{brms} object
}
\usage{
model_output_ts
}
\description{
Model Output - Two Source Model
Two source model output from \code{{brms}} model. .
}
\keyword{datasets}
