% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_output_tsar}
\alias{model_output_tsar}
\title{Model Output - Two Source Model - \eqn{\alpha_r}
Two source model with \eqn{\alpha_r} output from \code{{brms}} model. .}
\format{
\code{brms} object
}
\usage{
model_output_tsar
}
\description{
Model Output - Two Source Model - \eqn{\alpha_r}
Two source model with \eqn{\alpha_r} output from \code{{brms}} model. .
}
\keyword{datasets}
