% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_source_model.R
\name{one_source_model}
\alias{one_source_model}
\title{Bayesian model - One Source Trophic Position}
\usage{
one_source_model()
}
\description{
Estimate trophic position using a one source model derived from [Post 2002]()
}
