% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_source_priors_prams.R
\name{one_source_priors_prams}
\alias{one_source_priors_prams}
\title{Adjust Bayesian priors - One Source Trophic Position}
\usage{
one_source_priors_prams(n1 = NULL, n1_sigma = NULL)
}
\arguments{
\item{n1}{mean prior value for \eqn{\delta}\eqn{^1}\eqn{^5}N for baseline used in the model.
Defaults to 9.}

\item{n1_sigma}{variance for n1 priors that are \eqn{\delta}\eqn{^1}\eqn{^5}N for baseline used in the model.
Defaults to 1.}
}
\description{
Adjust priors for one source trophic position model derived from
\href{https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/0012-9658\%282002\%29083\%5B0703\%3AUSITET\%5D2.0.CO\%3B2}{Post 2002}.
}
\details{
This function allows the user to adjust the prior mean and variance for \eqn{\delta}\eqn{^1}\eqn{^5}N for baseline.
To be used with \code{one_source_priors} and \code{one_source_model}.

You can adjust the prior mean and variance for \eqn{\delta}\eqn{^1}\eqn{^5}N by supplying \code{n1} and \code{n1_sigma} with a numerical value.
}
\seealso{
\code{\link[=one_source_priors]{one_source_priors()}} and \code{\link[=one_source_model]{one_source_model()}}
}
