% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_priors}
\alias{set_priors}
\alias{default_priors}
\alias{supplied_priors}
\title{Set default priors}
\usage{
default_priors(
  model = c("one_source", "two_source", "two_source_ar", "two_source_arc")
)

supplied_priors(
  model = c("one_source", "two_source", "two_source_ar", "two_source_arc"),
  a = NULL,
  b = NULL,
  n1 = NULL,
  n1_sigma = NULL,
  n2 = NULL,
  n2_sigma = NULL,
  c1 = NULL,
  c1_sigma = NULL,
  c2 = NULL,
  c2_sigma = NULL,
  dn = NULL,
  dn_sigma = NULL,
  tp_lb = NULL,
  tp_ub = NULL,
  sigma_lb = NULL,
  sigma_ub = NULL
)
}
\arguments{
\item{model}{model type}
}
\description{
Set default priors
}
\keyword{internal}
