% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trps-package.R
\docType{package}
\name{trps-package}
\alias{trps}
\alias{trps-package}
\title{trps: Bayesian estimates of trophic position}
\description{
Bayesian estimates of trophic position leveraging 'brms' and using stable isotope data. Trophic position modeles are created using equations from Post (2002) \doi{10.1890/0012-9658(2002)083[0703:USITET]2.0.CO;2}, and Huevel et al. (2024) \doi{10.1139/cjfas-2024-0028}.
}
\author{
\strong{Maintainer}: Benjamin L. Hlina \email{benjamin.hlina@gmail.com}

}
\keyword{internal}
