% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_source_model.R
\name{two_source_model}
\alias{two_source_model}
\title{Bayesian model - Two Source Trophic Position}
\usage{
two_source_model()
}
\description{
Bayesian model for trophic position using a two source model derived from
[Post 2002](https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/0012-9658%282002%29083%5B0703%3AUSITET%5D2.0.CO%3B2).
}
\details{
\eqn{\delta}\eqn{^1}\eqn{^5}N = n1 * alpha + n2 * (1 - alpha) + dn * (tp - (\eqn{\lambda} * alpha + \eqn{\lambda} * (1 - alpha))

\eqn{\delta}\eqn{^1}\eqn{^5}N are values from the consumer,
n1 is \eqn{\delta}\eqn{^1}\eqn{^5}N  values of baseline 1,
\eqn{\Delta}N is the trophic discrimination factor for N (i.e., mean of 3.4),
tp is trophic position, and \eqn{\lambda} is the trophic level of baselines which are often a primary consumer (e.g., 2).

Posterior distributions of of n1, dn, and tp will be estimated.
}
