% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_source_model_ac.R
\name{two_source_model_ac}
\alias{two_source_model_ac}
\title{Bayesian model - Two Source Trophic Position with \eqn{\alpha_c}}
\usage{
two_source_model_ac(bp = FALSE, lambda = NULL)
}
\arguments{
\item{bp}{logical value that controls whether  priors are
supplied to the model for \eqn{\delta^{15}}N pelagic and benthic
baselines. Default is \code{FALSE} meaning the model will not be supplied priors,
however, the supplied \code{data.frame} needs values for
\eqn{\delta^{15}}N each baseline (\code{n1} and \code{n2}).}

\item{lambda}{numerical value, \code{1} or \code{2}, that controls whether one or
two lambdas are used. See details for equations and when to use \code{1} or \code{2}.
Defaults to \code{1}.}
}
\value{
returns model structure for two source model to be used in a
\code{brms()} call.
}
\description{
Bayesian model for trophic position using a two source model with \eqn{\alpha_c} derived from
\href{https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/0012-9658\%282002\%29083\%5B0703\%3AUSITET\%5D2.0.CO\%3B2}{Post 2002} and \href{https://cdnsciencepub.com/doi/10.1139/cjfas-2024-0028}{Heuvel et al. 2024}
}
\examples{
two_source_model_ac()

}
\seealso{
\code{\link[brms:brms-package]{brms::brms()}}
}
