% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_source_priors_ac.R
\name{two_source_priors_ac}
\alias{two_source_priors_ac}
\title{Bayesian priors - Two Source Trophic Position with \eqn{\alpha_c}}
\usage{
two_source_priors_ac(bp = FALSE)
}
\arguments{
\item{bp}{logical value that controls whether informed priors are
supplied to the model for both \eqn{\delta^{15}}N and
\eqn{\delta^{15}}C baselines. Default is \code{FALSE} meaning the model will
use uninformed priors, however, the supplied \code{data.frame} needs values
for both \eqn{\delta^{15}}N and \eqn{\delta^{15}}C baseline
(\code{c1}, \code{c2}, \code{n1}, and \code{n2}).}
}
\value{
returns priors for two source model to be used in a \code{brms()} call.
}
\description{
Create priors for trophic position using a two source model
with \eqn{\alpha_c} derived from \href{https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/0012-9658\%282002\%29083\%5B0703\%3AUSITET\%5D2.0.CO\%3B2}{Post 2002}
and \href{https://cdnsciencepub.com/doi/10.1139/cjfas-2024-0028}{Heuvel et al. 2024}.
}
\examples{
two_source_priors_ac()

}
\seealso{
\code{\link[brms:brms-package]{brms::brms()}}
}
