% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_functions}
\alias{error_functions}
\alias{check_column_names}
\alias{check_data_frame}
\alias{check_lambda}
\alias{check_logical}
\alias{check_numerical}
\title{Error functions}
\usage{
check_column_names(x, arg_name = NULL)

check_data_frame(x, arg_name = NULL)

check_lambda(x, arg_name = NULL)

check_logical(x, arg_name = NULL)

check_numerical(x, arg_name = NULL)
}
\arguments{
\item{x}{prior to check
@param arg_name the name of the argument to check.}

\item{arg_name}{the name of the argument to check.}
}
\description{
Error functions
}
\keyword{internal}
