% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_functions}
\alias{error_functions}
\alias{check_column_names}
\alias{check_data_frame}
\alias{check_logical}
\alias{check_numerical}
\title{Error functions}
\usage{
check_column_names(x)

check_data_frame(x)

check_logical(x)

check_numerical(x)
}
\arguments{
\item{x}{prior to check}
}
\description{
Error functions
}
\keyword{internal}
