% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_source_model.R
\name{one_source_model}
\alias{one_source_model}
\title{Bayesian model - One Source Trophic Position}
\usage{
one_source_model()
}
\description{
Bayesian model for trophic position using a one source model derived from [Post 2002](https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/0012-9658%282002%29083%5B0703%3AUSITET%5D2.0.CO%3B2)
}
\details{
\deqn{d15n ~ n1 + dn * (tp - lambda)}

where d15n are values from the consumer, n1 is d15n values of baseline 1,
\eqn{\Delta}is the trophic discrimination factor for N (i.e., mean of 3.4),
tp is trophic position, and lambda is the trophic level of baselines.

Posterior distributions of of n1, dn, and tp will be estimated.
}
