% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_priors}
\alias{set_priors}
\alias{default_priors}
\title{Set default priors}
\usage{
default_priors(
  model = c("one_source", "two_source", "two_source_ar", "two_source_arc")
)
}
\arguments{
\item{model}{model type}
}
\description{
Set default priors
}
\keyword{internal}
