% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trps-package.R
\docType{package}
\name{trps-package}
\alias{trps}
\alias{trps-package}
\title{trps: Bayesian Trophic Position Models using 'Stan'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bayesian trophic position models using 'stan' by leveraging 'brms' for stable isotope data. Trophic position models are derived by using equations from Post (2002) \doi{10.1890/0012-9658(2002)083[0703:USITET]2.0.CO;2}, Vander Zanden and Vadeboncoeur (2002) \doi{10.1890/0012-9658(2002)083[2152:FAIOBA]2.0.CO;2}, and Heuvel et al. (2024) \doi{10.1139/cjfas-2024-0028}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://benjaminhlina.github.io/trps/}
  \item \url{https://github.com/benjaminhlina/trps}
  \item Report bugs at \url{https://github.com/benjaminhlina/trps/issues}
}

}
\author{
\strong{Maintainer}: Benjamin L. Hlina \email{benjamin.hlina@gmail.com}

}
\keyword{internal}
