% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_source_model.R
\name{two_source_model}
\alias{two_source_model}
\title{Bayesian model - Two Source Trophic Position}
\usage{
two_source_model(bp = FALSE, lambda = NULL)
}
\arguments{
\item{bp}{logical value that controls whether  priors are
supplied to the model for \eqn{\delta}\eqn{^1}\eqn{^5}N pelagic and benthic
baselines. Default is \code{FALSE} meaning the model will not be supplied priors,
however, the supplied \code{data.frame} needs values for
\eqn{\delta}\eqn{^1}\eqn{^5}N each baseline (\code{n1} and \code{n2}).}

\item{lambda}{numerical value, \code{1} or \code{2}, that controls whether one or
two lambdas are used. See details for equations and when to use \code{1} or \code{2}.
Defaults to \code{1}.}
}
\value{
returns model structure for two source model to be used in a
\code{brms()} call.
}
\description{
Bayesian model for trophic position using a two source model derived from
\href{https://esajournals.onlinelibrary.wiley.com/doi/full/10.1890/0012-9658\%282002\%29083\%5B0703\%3AUSITET\%5D2.0.CO\%3B2}{Post 2002}.
}
\details{
When \code{lambda} is set to \code{1}

\deqn{\delta^{15}N = \Delta n \times (tp - \lambda) + n_1 \times \alpha + n_2 \times (1 - \alpha))}

or

When \code{lambda} is set to \code{2}

\deqn{\delta^{15}N = \Delta n \times (tp - (\lambda_1 \times \alpha + \lambda_2 \times (1 - \alpha))) + n_1 \times \alpha + n_2 \times (1 - \alpha))}

\eqn{\delta^{15}N} are values from the consumer,
\eqn{n_1} is \eqn{\delta^{15}N} values of baseline 1, \eqn{n_2} is
\eqn{\delta^{15}N} values of baseline 2,
\eqn{\Delta}N is the trophic discrimination factor for N (i.e., mean of \code{3.4}),
tp is trophic position, and \eqn{\lambda} is the trophic level of
baselines which are often a primary consumer (e.g., \code{2}).

The data supplied to \code{brms()} when using baselines at the same trophic level
(\code{lambda} argument set to \code{1}) needs to have the following variables, \code{d15n},
\code{n1}, \code{n2}, \code{l} (\eqn{\lambda}) which is usually \code{2 }. If using baselines that are at
different trophic levels (\code{lambda} argument set to \code{2}) the dataframe needs
to have \code{l1} and \code{l2} with a numerical value for each trophic level (e.g.,
\code{2} and \code{2.5}).
}
\examples{
two_source_model()

}
\seealso{
\code{\link[brms:brms-package]{brms::brms()}}
}
